Improve code by placing documentation on interfaces rather than in implementation
